Make URLs polymorphic
Refactor the Link type as a synonym for GLink, a polymorphic type whose
URLs are represented in different ways.

This is only a breaking change for weird uses like making Link an instance of a
new class (fixed by FlexibleInstances and TypeSynonymInstances) or
constructing a Link and immediately trying to print it without ever actually
inferring its concrete type.

Also bump to version 1.1.0, update stack.yaml, and set myself as maintainer.